25/07/18 upload:
abandoned the idea of adding filters, starting trying to implement a "send via mail" function. in order to check it, created a separate project "letstryagain" to check the "sending mail" function alone. unfortunately such function is still not working. ("letstryagain" project doesn't work, and in order to make the main "1st" project work the "mailmaybe.cpp" source file has to be removed, and then the program will execute normally, with only the functions to take pictures and save them as bitmap file in "Pictures" folder).